refactor: split layout_regression E2E test into 5 independent phase tests
- Split monolithic layout_regression.spec.ts (288 lines, 5 phases) into
  independent test files using loadData() API for state restoration
- Phase 1: Initial state & node addition (layout_phase1.spec.ts)
- Phase 2: Multiline text update (layout_phase2.spec.ts)
- Phase 3: Deep nesting & navigation (layout_phase3.spec.ts)
- Phase 4: Style changes & layout verification (layout_phase4.spec.ts)
- Phase 5: Layout mode switching (layout_phase5.spec.ts)
- Added JSON fixtures for state snapshots (tests/e2e/fixtures/)
- Updated Phase 4 screenshots for loadData-based viewport positioning
- All 14 E2E tests pass in parallel execution (~16s total)
