+++ tests/e2e/layout_phase1.spec.ts-snapshots/00-initial-state-chromium-linux.png
