+++ tests/e2e/layout_phase2.spec.ts-snapshots/02-multiline-update-chromium-linux.png
