+++ tests/e2e/layout_phase3.spec.ts-snapshots/03-children-added-chromium-linux.png
