+++ tests/e2e/layout_phase3.spec.ts-snapshots/05-after-navigation-chromium-linux.png
