+++ tests/e2e/layout_phase4.spec.ts-snapshots/07-after-style-nav-chromium-linux.png
