+++ tests/e2e/layout_phase5.spec.ts-snapshots/08-layout-left-chromium-linux.png
