+++ tests/e2e/layout_phase5.spec.ts-snapshots/09-layout-both-chromium-linux.png
